refactor(rendering): colocate SvgBoard under svg/components
Made-with: Cursor
