+++ src/rendering/svg/components/SvgBoard.stories.tsx
