+++ src/rendering/svg/components/SvgBoard.tsx
